% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_sl_library}
\alias{create_sl_library}
\title{Create SuperLearner library for TMLE}
\usage{
create_sl_library(
  include_glm = TRUE,
  include_gam = FALSE,
  include_rf = FALSE,
  include_gbm = FALSE,
  include_nnet = FALSE,
  include_mean = TRUE
)
}
\arguments{
\item{include_glm}{Logical to include glm}

\item{include_gam}{Logical to include gam}

\item{include_rf}{Logical to include random forest}

\item{include_gbm}{Logical to include gradient boosting}

\item{include_nnet}{Logical to include neural network}

\item{include_mean}{Logical to include mean (intercept-only model)}
}
\value{
A character vector of SuperLearner library names
}
\description{
Create SuperLearner library for TMLE
}
