% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package attachment function}
\usage{
.onAttach(libname, pkgname)
}
\description{
Ensures proper message and loading of namespace when the package is attached
with library() or require()
}
\keyword{internal}
