% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_variants.R
\name{format_tmle_results}
\alias{format_tmle_results}
\title{Format TMLE results}
\usage{
format_tmle_results(tmle_fit)
}
\arguments{
\item{tmle_fit}{A fitted tmle object}
}
\value{
A tibble with formatted results
}
\description{
Format TMLE results
}
