% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{run_multiple_tmle_variants}
\alias{run_multiple_tmle_variants}
\title{Run Multiple TMLE Variants on a Dataset}
\usage{
run_multiple_tmle_variants(
  data,
  outcome_var,
  treatment_var,
  covariates,
  tmle_variants = c("vanilla", "cvq_multiple", "fullcv_multiple", "doublecrossfit"),
  Q.SL.library = NULL,
  g.SL.library = NULL,
  V = 5,
  family = "binomial",
  num_repetitions = 100,
  sim_id_col = NULL,
  true_value = NULL,
  parallel_strategy = "none",
  n_cores = 2,
  slurm_options = list(),
  missing_handling = "complete_case",
  binary_outcome = TRUE,
  save_results = FALSE,
  results_dir = "./results",
  ...
)
}
\arguments{
\item{data}{A data frame containing analysis variables}

\item{outcome_var}{Name of the outcome variable}

\item{treatment_var}{Name of the treatment variable}

\item{covariates}{Vector of covariate names}

\item{tmle_variants}{Vector of TMLE variants to use (from 'vanilla', 'cvq', etc.)}

\item{Q.SL.library}{SuperLearner library for outcome model}

\item{g.SL.library}{SuperLearner library for treatment model}

\item{V}{Number of cross-validation folds}

\item{family}{Distribution family}

\item{num_repetitions}{Number of repetitions for multiple methods}

\item{sim_id_col}{Name of the column containing simulation IDs}

\item{true_value}{True parameter value(s) for coverage calculations}

\item{parallel_strategy}{Parallelization strategy}

\item{n_cores}{Number of cores to use for local parallelization}

\item{slurm_options}{List of SLURM options for SLURM parallelization}

\item{missing_handling}{Strategy for handling missing values}

\item{binary_outcome}{Logical indicating if outcome is binary}

\item{save_results}{Logical indicating whether to save results to file}

\item{results_dir}{Directory to save results}

\item{...}{Additional arguments passed to TMLE functions}
}
\value{
List containing combined TMLE analysis results from multiple variants
}
\description{
This function applies multiple TMLE variants to the same dataset and
combines the results for easy comparison.
}
