% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{split_simulation_data}
\alias{split_simulation_data}
\title{Split Data for Simulation Studies}
\usage{
split_simulation_data(data, sim_id_col = "sim_id")
}
\arguments{
\item{data}{A data frame containing simulation data}

\item{sim_id_col}{Name of the column containing simulation IDs}
}
\value{
A list of data frames, one for each simulation ID
}
\description{
Split Data for Simulation Studies
}
